dts: starfive: enhance SD card io Output Drive Strength
SD card gpio is generic gpio on visionfive2, the default output drive strength may be not enough, some sd card do not work.
So enhance the gpio output drive strength to max.
But on visionfive2, the sd card gpio is fixed 3.3V, there is no need to switch_voltage and not support uhs mode.

Signed-off-by: Jianlong Huang <jianlong.huang@starfivetech.com>